refactor: switch to ESM
BREAKING CHANGE:

Requires Node 14+

Also:
- chore: updates command-line-basics, devDep.
